requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
pillow

fastapi
uvicorn
python-multipart